--- a/src/glue/auth_handler.hpp
+++ b/src/glue/auth_handler.hpp
@@ -11,8 +11,6 @@
 
 #pragma once
 
-#include <regex>
-
 #include "auth/auth.hpp"
 #include "auth_global.hpp"
 #include "glue/auth.hpp"