Using ClangBuildAnalyzer to guide my efforts

|                          |  Before  | After    |
|--------------------------|----------|----------|
| Parsing (frontend)       | 1821.4 s | 1254.9 s |
| Codegen & opts (backend) |  462.5 s |  408.9 s |